fix: implement SphericalRotorConfigUtil with initialization and extraction functions
