+++ src/index.ts
@@ -2,3 +2,10 @@ export * from "./SphericalRotor.js";
 export * from "./AutoSphericalRotor.js";
 export * from "./RotorStopConfig.js";
 export * from "./SphericalRotorConfig.js";
+export * from "./SphericalRotorConfigUtil.js";
+
+import * as SphericalRotorConfigUtil from "./SphericalRotorConfigUtil.js";
+/**
+ * @deprecated Use named exports directly from "@masatomakino/threejs-spherical-rotor" instead.
+ */
+export { SphericalRotorConfigUtil };
